Move the log formatting function into `Logger`
